changed filenames to better project the contents that are in them
added more rules to the entrypoint.sh so all files are included

+++ entrypoint.sh
@@ -18,18 +18,45 @@ echo "repo name is" $GITHUB_REPOSITORY
 #setup here for making the build folder
 #check if the following files are present in ./github/workspace : [./src/data/contact.json, ./src/data/main_data.json]
 # if they are not present, then throw an error
-echo "checking if the following files are present in ./github/workspace : [./src/data/contacts.json, ./src/data/main_data.json]"
-if [ -f ./github/workspace/data/contacts.json ] && [ -f ./github/workspace/data/main_data.json ]
+echo "checking if the following files are present in ./github/workspace : [./src/data/contacts.json, ./src/data/main_data.json, ./src/data/project_crates.json, ./src/data/project_profiles.json, ./src/data/tabular_data.json]"
+if [ -f ./github/workspace/data/contacts.json ] && [ -f ./github/workspace/data/main_data.json ] && [ -f ./github/workspace/data/project_crates.json ] && [ -f ./github/workspace/data/project_profiles.json ] && [ -f ./github/workspace/data/tabular_data.json ];
 then
     echo "all files are present"
     #copy the files over into ./src/data
     echo "copying the files over into ./src/data"
     cp ./github/workspace/data/contacts.json ./src/data/contacts.json
     cp ./github/workspace/data/main_data.json ./src/data/main_data.json
+    cp ./github/workspace/data/project_crates.json ./src/data/project_crates.json
+    cp ./github/workspace/data/project_profiles.json ./src/data/project_profiles.json
+    cp ./github/workspace/data/tabular_data.json ./src/data/tabular_data.json
 else
-    echo "one or more files are missing"
+    echo "one or more files of the data folder are missing"
     exit 1
 fi
+
+#check if there is a img folder in ./github/workspace, if so copy it over to ./src/img recursively with force overwrite
+echo "checking if there is a img folder in ./github/workspace"
+if [ -d ./github/workspace/img ];
+then
+    echo "img folder is present"
+    echo "copying the img folder over to ./src/img"
+    cp -r ./github/workspace/img ./src/img
+else
+    echo "img folder is not present"
+fi
+
+#check if there is a readme.md file in ./github/workspace, if so thencopy it over , else throw an error
+echo "checking if there is a readme.md file in ./github/workspace"
+if [ -f ./github/workspace/readme.md ];
+then
+    echo "readme.md file is present"
+    echo "copying the readme.md file over to ./src/readme.md"
+    cp ./github/workspace/readme.md ./src/readme.md
+else
+    echo "readme.md file is not present"
+    exit 1
+fi
+
 echo "installing dependencies for building react app"
 npm install
 echo "npm run build"
